Inflection editing experience

+++ src/app/Http/Controllers/Resources/ContributionController.php
@@ -172,6 +172,10 @@ class ContributionController extends Controller
         $contribution = new Contribution;
         $contribution->account_id = $request->user()->id;
         $contribution->is_approved = null;
+        
+        if ($request->has('dependent_on_contribution_id')) {
+            $contribution->dependent_on_contribution_id = intval($request->input('dependent_on_contribution_id'));
+        }
 
         $this->saveContribution($contribution, $request);
 
@@ -258,11 +262,16 @@ class ContributionController extends Controller
             $this->contributionAlreadyApproved();
         }
 
-        ContributionControllerFactory::createController($contribution->type)->approve($contribution, $request);
+        if ($contribution->dependent_on !== null && ! $contribution->dependent_on->is_approved) {
+            abort(400, 'Contribution\'s dependencies are not approved.');
+        }
+
+        $id = ContributionControllerFactory::createController($contribution->type)->approve($contribution, $request);
 
         $contribution->is_approved = 1;
         $contribution->date_reviewed = Carbon::now();
         $contribution->reviewed_by_account_id = $request->user()->id;
+        $contribution->approved_as_entity_id = $id;
         $contribution->save();
 
         event(new ContributionApproved($contribution));
@@ -282,6 +291,11 @@ class ContributionController extends Controller
         $contribution = Contribution::findOrFail($id);
         $this->requestPermission($request, $contribution);
 
+        if ($contribution->is_approved) {
+            abort(400, 'You cannot delete an approved contribution.');
+        }
+
+        $contribution->dependencies()->delete();
         $contribution->delete();
 
         event(new ContributionDestroyed($contribution));
@@ -354,8 +368,9 @@ class ContributionController extends Controller
     protected function validateContributionRequest(Request $request)
     {
         $this->validate($request, [
-            'morph'           => 'required|string',
-            'contribution_id' => 'sometimes|required|numeric|exists:contributions,id'
+            'morph'                        => 'required|string',
+            'contribution_id'              => 'sometimes|numeric|exists:contributions,id',
+            'dependent_on_contribution_id' => 'sometimes|nullable|numeric|exists:contributions,id'
         ]);
     }
 
@@ -394,9 +409,13 @@ class ContributionController extends Controller
         $controller = ContributionControllerFactory::createController($request);
         $entity = $controller->populate($contribution, $request);
 
-        $contribution->type        = Morphs::getAlias($entity);
-        $contribution->language_id = $entity->language_id;
-        $contribution->notes       = $request->has('notes') 
+        $type = Morphs::getAlias($entity);
+        if ($type === null) {
+            abort(400, 'Unrecognised type of contribution.');
+        }
+
+        $contribution->type  = $type;
+        $contribution->notes = $request->has('notes') 
             ? $request->input('notes') 
             : null;
